refactor(config): enumerate the path fields once
Anchoring and expansion each carried their own list of which fields hold
paths. A path field added to the schema would have joined one list and
not the other, silently: expanded but never anchored is issue #732 back
for that field alone, anchored but never expanded hands a literal `~` to
the filesystem. Neither is a compile error, and neither is loud at
runtime.

Both now walk `paths::path_fields_mut`, which destructures
`LanguageSettings` exhaustively — a new field fails to compile until
someone decides whether it is a path. Verified by adding one and
watching E0027 fire.

The walk carries the name ordering that expansion needed anyway for
deterministic error reporting, and anchoring moves here with it:
`expand.rs` is about expanding strings, and this was the one item in it
that knew what a settings struct looks like.

Review findings: design 1/2/4, tests 6.
