+++ src/config/expand.rs
@@ -1,4 +1,3 @@
-use path_clean::PathClean;
 use std::fmt;
 use std::path::{Path, PathBuf};
 use std::sync::OnceLock;
@@ -156,115 +155,6 @@ pub(super) fn expand_path(
     }
 }
 
-/// Whether a configured path already says where it lives, so anchoring must
-/// leave it alone.
-///
-/// Three cases, and the last two are why this is a syntactic test rather than
-/// `Path::is_absolute` alone — they are not absolute *yet*:
-/// - absolute by the platform's own rule (`/usr/share`, and on Windows a path
-///   carrying a drive prefix or root);
-/// - `~`-led, which expansion turns into an absolute path — including `~user`,
-///   which expansion deliberately passes through unchanged;
-/// - `$`-led, which expansion resolves to wherever the variable points. Joining
-///   a base onto that syntax would corrupt it before expansion reads it, and
-///   deciding otherwise would mean expanding here — which would consume the
-///   `${KAKEHASHI_DATA_DIR}` template the defaults layer depends on.
-///
-/// A bare `$$literal` is skipped too, even though it expands to the *relative*
-/// literal `$literal`. Answering otherwise would require distinguishing the
-/// escape from a variable reference, i.e. parsing the expansion syntax here.
-/// To place any of these under the source directory, lead with `./`.
-fn carries_its_own_base(path: &str) -> bool {
-    Path::new(path).is_absolute()
-        // On Windows a POSIX-rooted value is not `is_absolute`, but it is still
-        // rooted and joining a base onto it would be wrong.
-        || path.starts_with('/')
-        || path.starts_with('~')
-        || path.starts_with('$')
-}
-
-/// Rewrite every relative path field in `settings` to sit under `base`.
-///
-/// Called once per configuration layer, while the directory that layer came
-/// from is still known — after the layers merge, a surviving `./queries` no
-/// longer says which file asked for it, and the only base left is the server's
-/// working directory, which belongs to whoever launched the editor.
-///
-/// This runs *before* expansion and deliberately stays syntactic; see
-/// [`carries_its_own_base`] for which values it declines to touch and why.
-///
-/// Idempotent, and load-bearingly so: anchoring yields an absolute path, which
-/// `carries_its_own_base` then skips. That is what lets `didChangeConfiguration`
-/// merge a freshly anchored layer onto already-anchored stored settings, and the
-/// post-install reload re-derive settings, without re-basing anything.
-///
-/// `base` is `None` for layers with no source directory (the programmed
-/// defaults), where every value is left untouched.
-pub(crate) fn anchor_settings_paths(
-    settings: &mut crate::config::RawWorkspaceSettings,
-    base: Option<&Path>,
-) {
-    let Some(base) = base else {
-        return;
-    };
-
-    // Anchoring must not *introduce* expansion syntax either. The expansion pass
-    // reads `$VAR` anywhere in the value, not just at the front, so a base
-    // directory that itself contains a `$` — `/data/$USER/proj`, a directory
-    // literally named `a$b` — would be read as a variable reference in every
-    // path this layer anchors. That fails the whole configuration when the name
-    // is undefined, and silently resolves somewhere else when it happens to be
-    // defined. `$$` is the documented escape for a literal `$`; it is what the
-    // one expansion pass will turn back into the directory's real name, and
-    // `clean` leaves it alone because it folds components, not characters.
-    let base = PathBuf::from(base.to_string_lossy().replace('$', "$$"));
-
-    let anchor = |path: &mut String| {
-        if carries_its_own_base(path) {
-            return;
-        }
-        let joined = base.join(&*path);
-        // Folding `./` and `../` away keeps the stored value readable, but the
-        // fold is lexical: `..` pops whatever component precedes it, and before
-        // expansion that component may be a variable. Popping `$VAR` resolves
-        // somewhere else the moment the variable holds more than one component,
-        // and it takes any undefined-variable error down with it — expansion can
-        // only reject a variable it still sees. So a value carrying both is left
-        // unfolded for the kernel to resolve. Either alone is safe: with no
-        // `..` there is nothing to pop, and with no variable there is nothing
-        // whose value the fold could get wrong.
-        //
-        // The `.` segments that joining introduces are dropped either way —
-        // a `.` pops nothing, so removing it cannot move the path.
-        *path = if path.contains('$') && path.contains("..") {
-            joined
-                .components()
-                .filter(|component| !matches!(component, std::path::Component::CurDir))
-                .collect::<PathBuf>()
-                .to_string_lossy()
-                .into_owned()
-        } else {
-            joined.clean().to_string_lossy().into_owned()
-        };
-    };
-
-    if let Some(search_paths) = settings.search_paths.as_mut() {
-        for path in search_paths {
-            anchor(path);
-        }
-    }
-    for language in settings.languages.values_mut() {
-        if let Some(parser) = language.parser.as_mut() {
-            anchor(parser);
-        }
-        if let Some(queries) = language.queries.as_mut() {
-            for query in queries {
-                anchor(&mut query.path);
-            }
-        }
-    }
-}
-
 /// Wrap an env lookup function to provide fallback values for known `KAKEHASHI_`
 /// environment variables when they are undefined.
 ///
@@ -313,219 +203,6 @@ pub(crate) fn make_env(vars: &[(&str, &str)]) -> impl Fn(&str) -> Option<String>
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    /// Build settings whose every path-valued field carries `value`, so one
-    /// assertion covers `searchPaths`, `parser`, and `queries[].path` alike.
-    fn settings_with_path(value: &str) -> crate::config::RawWorkspaceSettings {
-        crate::config::RawWorkspaceSettings {
-            search_paths: Some(vec![value.to_string()]),
-            languages: [(
-                "lua".to_string(),
-                crate::config::LanguageSettings {
-                    parser: Some(value.to_string()),
-                    queries: Some(vec![crate::config::settings::QueryItem {
-                        path: value.to_string(),
-                        kind: None,
-                    }]),
-                    ..Default::default()
-                },
-            )]
-            .into(),
-            ..Default::default()
-        }
-    }
-
-    /// Every path field of `settings`, so a rule cannot be asserted for one
-    /// field and quietly missed for another.
-    fn all_paths(settings: &crate::config::RawWorkspaceSettings) -> Vec<String> {
-        let language = &settings.languages["lua"];
-        let mut paths = settings.search_paths.clone().unwrap_or_default();
-        paths.extend(language.parser.clone());
-        paths.extend(language.queries.iter().flatten().map(|q| q.path.clone()));
-        paths
-    }
-
-    fn anchor(value: &str, base: Option<&Path>) -> String {
-        let mut settings = settings_with_path(value);
-        anchor_settings_paths(&mut settings, base);
-        let paths = all_paths(&settings);
-        assert_eq!(paths.len(), 3, "every path field should be visited");
-        assert!(
-            paths.iter().all(|path| path == &paths[0]),
-            "every path field should follow the same rule: {paths:?}"
-        );
-        paths[0].clone()
-    }
-
-    #[test]
-    fn relative_paths_are_anchored_to_the_base() {
-        assert_eq!(
-            anchor("./queries/highlights.scm", Some(Path::new("/workspace"))),
-            "/workspace/queries/highlights.scm"
-        );
-        assert_eq!(
-            anchor("queries/highlights.scm", Some(Path::new("/workspace"))),
-            "/workspace/queries/highlights.scm"
-        );
-    }
-
-    #[test]
-    fn parent_traversal_is_normalized_away() {
-        assert_eq!(
-            anchor("../shared/parsers", Some(Path::new("/workspace/project"))),
-            "/workspace/shared/parsers"
-        );
-    }
-
-    /// Anchoring runs before expansion, so it must not disturb the syntax the
-    /// expansion pass reads — including the `$$` escape for a literal dollar.
-    #[test]
-    fn variables_survive_anchoring_unexpanded() {
-        assert_eq!(
-            anchor("./queries/$LANG.scm", Some(Path::new("/workspace"))),
-            "/workspace/queries/$LANG.scm"
-        );
-        assert_eq!(
-            anchor("./$$literal", Some(Path::new("/workspace"))),
-            "/workspace/$$literal"
-        );
-    }
-
-    /// A base directory may legally contain a `$`, and the expansion pass reads
-    /// `$VAR` anywhere in a value — not just at the front. Anchoring therefore
-    /// has to escape what it prepends, or every path in that layer would be
-    /// rejected as an undefined variable, or worse, silently resolved to
-    /// somewhere else when the name happens to be defined.
-    #[test]
-    fn a_base_containing_a_dollar_survives_the_expansion_pass() {
-        let anchored = anchor("./queries", Some(Path::new("/work/a$b")));
-        assert_eq!(anchored, "/work/a$$b/queries");
-
-        let env = make_env(&[("b", "SHOULD NOT BE USED")]);
-        assert_eq!(
-            expand_path(&anchored, None, &env).expect("the escaped base must expand cleanly"),
-            "/work/a$b/queries",
-            "expansion must give back the directory's real name"
-        );
-    }
-
-    /// Folding `..` lexically would pop the *unexpanded* component in front of
-    /// it, which is not the same path and can be no path at all: the variable
-    /// disappears along with the undefined-variable error the expansion pass
-    /// owes the user, and a variable holding more than one component resolves
-    /// somewhere else entirely. A value carrying a variable is therefore left
-    /// unfolded for the filesystem to resolve.
-    #[test]
-    fn parent_traversal_after_a_variable_is_left_for_the_filesystem() {
-        assert_eq!(
-            anchor("./a/$VAR/../b", Some(Path::new("/base"))),
-            "/base/a/$VAR/../b"
-        );
-        assert_eq!(
-            anchor("./a/$UNDEFINED/../b", Some(Path::new("/base"))),
-            "/base/a/$UNDEFINED/../b",
-            "the undefined variable must still reach expansion, which rejects it"
-        );
-    }
-
-    /// A value that expansion turns into an absolute path is left alone: its
-    /// author already said where it lives, and joining a base onto a leading
-    /// `~` or `$` would corrupt the syntax before expansion ever sees it.
-    #[test]
-    fn absolute_and_expandable_prefixes_are_not_anchored() {
-        let base = Some(Path::new("/workspace"));
-        assert_eq!(
-            anchor("/opt/kakehashi/runtime", base),
-            "/opt/kakehashi/runtime"
-        );
-        assert_eq!(anchor("~/parsers/lua.so", base), "~/parsers/lua.so");
-        assert_eq!(anchor("$KAKEHASHI_DATA_DIR", base), "$KAKEHASHI_DATA_DIR");
-        assert_eq!(
-            anchor("${KAKEHASHI_DATA_DIR}", base),
-            "${KAKEHASHI_DATA_DIR}"
-        );
-    }
-
-    /// The opt-out rule, stated once so `docs/README.md` has something to match.
-    #[test]
-    fn carries_its_own_base_covers_every_opt_out_form() {
-        for path in [
-            "/usr/share/kakehashi",
-            "~/parsers",
-            "~",
-            // Expansion passes `~user` through unchanged, so anchoring it would
-            // produce a directory literally named `~bob` under the base.
-            "~bob/parsers",
-            "$KAKEHASHI_DATA_DIR",
-            "${KAKEHASHI_DATA_DIR}/queries",
-            // Skipped despite expanding to the relative literal `$literal`:
-            // telling it apart from a variable means parsing the syntax here.
-            "$$literal",
-        ] {
-            assert!(carries_its_own_base(path), "{path} should opt out");
-        }
-
-        for path in ["queries/x.scm", "./queries/x.scm", "../shared", ".", ""] {
-            assert!(!carries_its_own_base(path), "{path} should be anchored");
-        }
-    }
-
-    /// A Windows-absolute value must opt out on Windows, where neither a drive
-    /// prefix nor a bare root starts with `/`.
-    #[cfg(windows)]
-    #[test]
-    fn windows_absolute_paths_opt_out() {
-        for path in [r"C:\parsers\lua.dll", r"\\server\share\lua.dll", r"\rooted"] {
-            assert!(carries_its_own_base(path), "{path} should opt out");
-        }
-    }
-
-    /// One array, three rules: the relative entry moves, the two that carry
-    /// their own base do not. The single-value helper above cannot express
-    /// this, and a layer's `searchPaths` routinely mixes the forms.
-    #[test]
-    fn one_array_may_mix_anchored_and_untouched_entries() {
-        let mut settings = crate::config::RawWorkspaceSettings {
-            search_paths: Some(vec![
-                "./runtime".into(),
-                "/opt/kakehashi".into(),
-                "${KAKEHASHI_DATA_DIR}".into(),
-            ]),
-            ..Default::default()
-        };
-
-        anchor_settings_paths(&mut settings, Some(Path::new("/workspace")));
-
-        assert_eq!(
-            settings.search_paths,
-            Some(vec![
-                "/workspace/runtime".to_string(),
-                "/opt/kakehashi".to_string(),
-                "${KAKEHASHI_DATA_DIR}".to_string(),
-            ])
-        );
-    }
-
-    /// Degenerate spellings of "here". Both name the source directory, which is
-    /// what anchoring resolves them to — worth pinning because `expand_path`
-    /// leaves an empty string empty, so the two passes disagree by design.
-    #[test]
-    fn empty_and_dot_resolve_to_the_base_itself() {
-        assert_eq!(anchor("", Some(Path::new("/workspace"))), "/workspace");
-        assert_eq!(anchor(".", Some(Path::new("/workspace"))), "/workspace");
-    }
-
-    /// The programmed defaults have no source directory. Without a base every
-    /// value must survive verbatim, or `${KAKEHASHI_DATA_DIR}` would stop
-    /// reaching the expansion pass that gives it a platform default.
-    #[test]
-    fn no_base_leaves_every_path_untouched() {
-        assert_eq!(anchor("./runtime", None), "./runtime");
-        assert_eq!(
-            anchor("${KAKEHASHI_DATA_DIR}", None),
-            "${KAKEHASHI_DATA_DIR}"
-        );
-    }
 
     #[test]
     fn expand_dollar_var() {
